+++ packages/vibrant-core/src/lib/Link/Link.tsx
@@ -1,16 +1,29 @@
+import { isDefined } from '@vibrant-ui/utils';
 import { Box } from '../Box';
 import { useConfig } from '../ConfigProvider';
 import { withLinkVariation } from './LinkProps';
 
 export const Link = withLinkVariation(
-  ({ innerRef, isExternal, onFocusIn, onFocusOut, onHoverIn, onHoverOut, onPressIn, onPressOut, testId, ...props }) => {
+  ({
+    innerRef,
+    isExternal,
+    onFocusIn,
+    onFocusOut,
+    onHoverIn,
+    onHoverOut,
+    onPressIn,
+    onPressOut,
+    testId,
+    hitSlop: _,
+    ...props
+  }) => {
     const {
       dependencies: { link },
     } = useConfig();
 
     return (
       <Box
-        as="a"
+        {...(isDefined(link) ? {} : { as: 'a' })}
         data-testid={testId}
         base={link}
         ref={innerRef}