+++ packages/vibrant-core/src/lib/Link/LinkProps.ts
@@ -14,7 +14,7 @@ import type {
 } from '../props';
 import { withVariation } from '../withVariation';
 
-type LinkProps = {
+export type LinkProps = {
   testId?: string;
   ref?: any;
   href: string;