+++ packages/vibrant-core/src/lib/PressableBox/index.ts
@@ -1,1 +1,2 @@
 export { PressableBox } from './PressableBox';
+export type { PressableBoxProps } from './PressableBoxProps';