+++ packages/vibrant-components/CHANGELOG.md
@@ -2,6 +2,39 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.53.6](https://github.com/pedaling/opensource/compare/vibrant-components-0.53.5...vibrant-components-0.53.6) (2023-04-05)
+
+### [0.53.5](https://github.com/pedaling/opensource/compare/vibrant-components-0.53.4...vibrant-components-0.53.5) (2023-04-05)
+
+### [0.53.4](https://github.com/pedaling/opensource/compare/vibrant-components-0.53.3...vibrant-components-0.53.4) (2023-04-05)
+
+### [0.53.3](https://github.com/pedaling/opensource/compare/vibrant-components-0.53.2...vibrant-components-0.53.3) (2023-04-05)
+
+## [0.53.0](https://github.com/pedaling/opensource/compare/vibrant-components-0.52.3...vibrant-components-0.53.0) (2023-04-04)
+
+
+### Features
+
+* Avatar 컴포넌트를 추가한다 ([#655](https://github.com/pedaling/opensource/issues/655)) ([d385636](https://github.com/pedaling/opensource/commit/d385636bd4ba5a08a2f4752cb5365e1548dd05b4))
+
+### [0.52.3](https://github.com/pedaling/opensource/compare/vibrant-components-0.52.2...vibrant-components-0.52.3) (2023-03-31)
+
+### [0.52.2](https://github.com/pedaling/opensource/compare/vibrant-components-0.52.1...vibrant-components-0.52.2) (2023-03-30)
+
+
+### Bug Fixes
+
+* don't pass as="a" in Pressable ([#652](https://github.com/pedaling/opensource/issues/652)) ([ec977a2](https://github.com/pedaling/opensource/commit/ec977a23ecc13a57c50811dc83b37ad89162d1c9))
+
+### [0.52.1](https://github.com/pedaling/opensource/compare/vibrant-components-0.52.0...vibrant-components-0.52.1) (2023-03-30)
+
+## [0.52.0](https://github.com/pedaling/opensource/compare/vibrant-components-0.51.4...vibrant-components-0.52.0) (2023-03-30)
+
+
+### Features
+
+* Contained/Outlined/Ghost/IconButton, FilterChip 컴포넌트에 href 속성을 추가한다 ([#650](https://github.com/pedaling/opensource/issues/650)) ([c922324](https://github.com/pedaling/opensource/commit/c922324de8db4e22c5d4dc091cfb9a2dd8726d17))
+
 ### [0.51.4](https://github.com/pedaling/opensource/compare/vibrant-components-0.51.3...vibrant-components-0.51.4) (2023-03-27)
 
 ### [0.51.3](https://github.com/pedaling/opensource/compare/vibrant-components-0.51.2...vibrant-components-0.51.3) (2023-03-26)